update to ABI 3 fixes
- remove `renderfp` field from `NcOptions`.
- change to static inline functions: `notcurses_[top|bottom]`, `nccells_[double|rounded]_box`, `ncdirect_canopen_images`, `ncstrwidth`, `ncplane_putstr_aligned`, `ncplane_putstr_yx`, `ncplane_putstr_stained`, `notcurses_can*`, `notcurses_render`.
- remove: `ncplane_highgradient`, `ncplane_highgradient_sized`, `ncvisual_subtitle`, `ncplane_pixelgeom`, `notcurses_mouse_disable`, `notcurses_mouse_enable`, `ncvisual_blitter_geom`, `NcBlitterGeometry`, `ncvisual_render`, `nctablet_ncplane`, `nccell_width`, `getc`.
- add `ncplane_gradient2x1`, `notcurses_capabilities`.
- rename `ncmetric` to `ncnmetric`, and update function.
- rename `ncdirect_getc_[nblock|blocking]` to `ncdirect_get_[nblock|blocking]`, and `ncdirect_getc_[nblock|blocking]` to `ncdirect_get_[nblock|blocking]`.
- fix ncmetric constants.
- fix expected i32/u32 types.
- fix examples.
- enable `.set_size_t_is_usize` option in bindgen.

+++ src/visual/methods/options.rs
@@ -49,11 +49,11 @@ impl NcVisualOptions {
             y: y as i32,
             x: x as i32,
             // origin of rendered section
-            begy: beg_y as i32,
-            begx: beg_x as i32,
+            begy: beg_y,
+            begx: beg_x,
             // size of rendered section
-            leny: len_y as i32,
-            lenx: len_x as i32,
+            leny: len_y,
+            lenx: len_x,
             // glyph set to use
             blitter,
             // bitmask over NCVISUAL_OPTION_*
@@ -86,11 +86,11 @@ impl NcVisualOptions {
             y: y as i32,
             x: x as i32,
             // origin of rendered section
-            begy: beg_y as i32,
-            begx: beg_x as i32,
+            begy: beg_y,
+            begx: beg_x,
             // size of rendered section
-            leny: len_y as i32,
-            lenx: len_x as i32,
+            leny: len_y,
+            lenx: len_x,
             // glyph set to use
             blitter,
             // bitmask over NCVISUAL_OPTION_*
